Worked on NCIATWP-672 ( Summary Data Section ) and NCIATWP-732 ( Reset Study )
Basically I worked on creating the python code for accessing the R program.  I have tested it and it seems work , but it needs to be integrated into the web layer ( an ajax javascript call )

Due to reset being an eaiser function to write., I implemented that first since it will make function to access the python code eaiser to do.